UI Handler
The UI Handler class has been created and implemented. It is designed to prevent the need for the develop to handle or forward the touch events to each individual piece of UI in every scene. This overall will reduce the amount of code in each scene and the complexity that the developer has to worry about.

UI is added to the UIHandler when a touch listener is added to it, and it is removed when the scene is changed (handled by the SceneManager)