New RenderObject Method
Added a new render object method for rendering 2D objects with a custom model matrix.

Fixed a bug in the dropdown that gives the wrong selected item ID when clicking no items.

Fixed a bug in the text where it would generate with the wrong colour when using setText method