Fixed Touch Event Issues
The touch method in a scene that is called by the CrispinMobile engine would be called from a different thread to the rest of the scene. This means that certain actions such as creating textures where not completed on the graphics thread and applications would run into unexpected errors or have to implement a fix themselves. Now in the touch event you do not have to worry about thread safety as it is running on the same thread as the rest of a scene. This was accomplished by adding a blocking queue to the scene manager that could be fed the motion event data. The queue was then read in the main loop (in SceneManager) just after the update for each scene.

Other changes such as flipping the touch Y axis and the ability to scale 2D vectors.